feat: pie chart ux ui changes
