+++ packages/frontend/src/components/VisualizationConfigs/PieChartConfig/Display/Display.tsx
@@ -2,10 +2,17 @@ import {
     PieChartLegendPositions,
     type PieChartLegendPosition,
 } from '@lightdash/common';
-import { Collapse, SegmentedControl, Stack, Switch, Text } from '@mantine/core';
+import {
+    Collapse,
+    Group,
+    SegmentedControl,
+    Stack,
+    Switch,
+} from '@mantine/core';
 import React from 'react';
 import { isPieVisualizationConfig } from '../../../LightdashVisualization/VisualizationConfigPie';
 import { useVisualizationContext } from '../../../LightdashVisualization/VisualizationProvider';
+import { Config } from '../../common/Config';
 
 export const Display: React.FC = () => {
     const { visualizationConfig } = useVisualizationContext();
@@ -21,30 +28,30 @@ export const Display: React.FC = () => {
 
     return (
         <Stack>
-            <Switch
-                label="Show legend"
-                checked={showLegend}
-                onChange={toggleShowLegend}
-            />
+            <Config>
+                <Group>
+                    <Config.Label>Show legend</Config.Label>
+                    <Switch checked={showLegend} onChange={toggleShowLegend} />
+                </Group>
+            </Config>
 
             <Collapse in={showLegend}>
-                <Text fw={600}>Orientation</Text>
-                <SegmentedControl
-                    name="orient"
-                    color="blue"
-                    size="sm"
-                    fullWidth
-                    value={legendPosition}
-                    onChange={(val: PieChartLegendPosition) =>
-                        legendPositionChange(val)
-                    }
-                    data={Object.entries(PieChartLegendPositions).map(
-                        ([position, label]) => ({
-                            label,
-                            value: position,
-                        }),
-                    )}
-                />
+                <Group spacing="xs">
+                    <Config.SubLabel>Orientation</Config.SubLabel>
+                    <SegmentedControl
+                        name="orient"
+                        value={legendPosition}
+                        onChange={(val: PieChartLegendPosition) =>
+                            legendPositionChange(val)
+                        }
+                        data={Object.entries(PieChartLegendPositions).map(
+                            ([position, label]) => ({
+                                label,
+                                value: position,
+                            }),
+                        )}
+                    />
+                </Group>
             </Collapse>
         </Stack>
     );
